# docs/openapi.yaml
...
openapi: '3.1.0'
info:
  title: SomniBot Dashboard API
  version: '1.0.0'
  description: |
    Internal API for the SomniBot dashboard. All routes require guild-owner
    authentication via Discord OAuth session unless noted otherwise.

    **Authentication**: Discord OAuth2 session cookie (managed by NextAuth).
    Most endpoints also require `X-Guild-ID` header or derive guild from session.
  contact:
    email: heyimdionysus@gmail.com
servers:
  - url: '{dashboardUrl}'
    description: Dashboard instance
    variables:
      dashboardUrl:
        default: 'http://localhost:3000'
security:
  - sessionCookie: []
components:
  securitySchemes:
    sessionCookie:
      type: apiKey
      in: cookie
      name: next-auth.session-token
      description: NextAuth session cookie from Discord OAuth
...
paths:
...
  /api/orders/{id}/refund:
    post:
      summary: 'POST /api/orders/{id}/refund'
      description: 'Drive one durable, attempt-keyed owner refund. Every non-2xx/422 outcome is safely retryable with the same order id; the attempt is resumed, never duplicated.'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                reason:
                  type: string
                  maxLength: 255
                  default: 'Admin refund'
                revoke_entitlements:
                  const: true
                  default: true
                  description: 'A full owner refund always revokes the purchased access; false is rejected.'
      responses:
        '200':
          description: 'Completed — { success: true, status: completed }'
        '202':
          description: 'REFUND_PENDING — PayPal is still processing; access unchanged'
        '400':
          description: 'INVALID_ORDER_ID or validation failure'
        '409':
          description: 'ORDER_NOT_REFUNDABLE — order cannot be refunded from the owner order list'
        '422':
          description: 'PROVIDER_FAILED or PROVIDER_CANCELLED — PayPal terminally rejected the refund; customer access remains active'
        '500':
          description: 'LOCAL_FINALIZATION_PENDING (provider refunded, access cleanup pending — retry to finish), REFUND_STATE_SAVE_FAILED, or REFUND_PREPARATION_FAILED'
        '502':
          description: 'PROVIDER_REQUEST_UNCONFIRMED — PayPal status unconfirmed; retry safely'
...